requirements.txt
torch
torchvision
pandas
scikit-learn
numpy
pillow
streamlit
